requirements.txt
matplotlib==3.2.2
extra_streamlit_components==0.1.43
anastruct==1.2.0
streamlit==0.88.0
Pillow==8.3.2
